+++ node/src/main/java/CordaCaplet.java
@@ -112,8 +112,6 @@ public class CordaCaplet extends Capsule {
                 // If it fails, just return the existing class path. The main Corda jar will detect the error and fail gracefully.
                 return cp;
             }
-            // Add additional directories of JARs to the classpath (at the end), e.g., for JDBC drivers.
-            augmentClasspath((List<Path>) cp, cordappsDir);
             try {
                 List<String> jarDirs = nodeConfig.getStringList("jarDirs");
                 log(LOG_VERBOSE, "Configured JAR directories = " + jarDirs);