+++ docs/source/changelog.rst
@@ -9,6 +9,11 @@ Unreleased
 
 * Added auto-acceptance of network parameters for network updates. This behaviour is available for a subset of the network parameters
   and is configurable via the node config. See :doc:`network-map` for more information.
+  
+* Deprecated `SerializationContext.withAttachmentsClassLoader`. This functionality has always been disabled by flags
+and there is no reason for a CorDapp developer to use it. It is just an internal implementation detail of Corda.
+
+* Deprecated the `LedgerTransaction` constructor. No client code should call it directly. LedgerTransactions can be created from WireTransactions if required.
 
 * Introduced new optional network bootstrapper command line options (--register-package-owner, --unregister-package-owner)
   to register/unregister a java package namespace with an associated owner in the network parameter packageOwnership whitelist.